% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grammar.R
\name{grammar}
\alias{grammar}
\title{Spell Correcting Function}
\usage{
grammar(corpus)
}
\arguments{
\item{corpus}{Corpus object.}
}
\value{
data.table  Data table containing the original and grammar 
cleaned corpus text
}
\description{
This function will stem and correct all spelling in the corpus using 
the TreeTag tool, the koRpus package and the stringr package. The 
function will return a data.table.
}
\examples{
grammar(text.corpus)
}
\author{
Sophie C Haynes, \email{thesophienator96@gmail.com}
}
\keyword{data.table::data.table()}
\keyword{korPus::treetag()}
\keyword{korRus::taggedText()}
\keyword{nlp}
\keyword{spelling}
\keyword{stemming}
\keyword{stringr::str_replace_all()}
\keyword{treetag}
