% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punctStrip.R
\name{punctStrip}
\alias{punctStrip}
\title{Corpus Punctuation Stripping Function}
\usage{
punctStrip(corpus)
}
\arguments{
\item{corpus}{Corpus object.}
}
\description{
This function strips punctuation from a corpus object.
}
\examples{
punctStrip(text.corpus)
}
\author{
Sophie C Haynes, \email{thesophienator96@gmail.com}
}
\keyword{punctuation}
\keyword{textmine}
\keyword{tm}
\keyword{tm::tm_map()}
